Store bucket testing selection with EventHandler
Converted the Silex ServiceProvider BucketSelectionServiceProvider into
an EventHandler. This will ensure smoother transition to Symfony.
